requirements.txt
langchain
langchain-community
langchain-google-genai
streamlit
faiss-cpu
python-dotenv
google-generativeai
streamlit-mermaid